# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 61a8d1825db62a1b83fc156c
  title: schoolmanagement
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/course/{id}":
    delete:
      x-skaffolder-id: 61a8d5eb5db62a1b83fc1622
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61a8d42d5db62a1b83fc15ef
      x-skaffolder-resource: course
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61a8d5eb5db62a1b83fc1616
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61a8d42d5db62a1b83fc15ef
      x-skaffolder-resource: course
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/course"
    post:
      x-skaffolder-id: 61a8d5eb5db62a1b83fc1619
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61a8d42d5db62a1b83fc15ef
      x-skaffolder-resource: course
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/course"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/course"
...
  "/exam/{id}":
    delete:
      x-skaffolder-id: 61a8d5fc5db62a1b83fc162b
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61a8d5fc5db62a1b83fc1628
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
    post:
      x-skaffolder-id: 61a8d5fc5db62a1b83fc1626
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exam"
  "/exam/findBy_course/{key}":
    get:
      x-skaffolder-id: 61a8d5eb5db62a1b83fc161b
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_student/{key}":
    get:
      x-skaffolder-id: 61a8d5a75db62a1b83fc1602
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 61a8d5d85db62a1b83fc160f
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/{id}/validate":
    post:
      x-skaffolder-id: 61a8d7265db62a1b83fc1630
      x-skaffolder-name: validate
      x-skaffolder-id-resource: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: this API is to validate the exam
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /{id}/validate
      x-skaffolder-roles:
      summary: this API is to validate the exam
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: id of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    course:
      x-skaffolder-id: 61a8d42d5db62a1b83fc15ef
      x-skaffolder-id-db: 61a8d1825db62a1b83fc156d
      x-skaffolder-id-entity: 61a8d42d5db62a1b83fc15ed
      x-skaffolder-url: /course
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 61a8d42d5db62a1b83fc15ee
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    exam:
      x-skaffolder-id: 61a8d46e5db62a1b83fc15f5
      x-skaffolder-id-db: 61a8d1825db62a1b83fc156d
      x-skaffolder-id-entity: 61a8d46d5db62a1b83fc15f2
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 61a8d46d5db62a1b83fc15f4
          x-skaffolder-type: String
        score:
          type: number
          x-skaffolder-id-attr: 61a8d46d5db62a1b83fc15f3
          x-skaffolder-type: Number
        valid:
          type: boolean
          x-skaffolder-id-attr: 61a8d8b65db62a1b83fc1633
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 61a8d4f75db62a1b83fc15f8
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61a8d46d5db62a1b83fc15f2
          x-skaffolder-ent2: 61a8d42d5db62a1b83fc15ed
        _student:
          x-skaffolder-id: 61a8d4f75db62a1b83fc15f7
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61a8d46d5db62a1b83fc15f2
          x-skaffolder-ent2: 61a8d3b65db62a1b83fc15e1
        _teacher:
          x-skaffolder-id: 61a8d4f75db62a1b83fc15f9
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61a8d46d5db62a1b83fc15f2
          x-skaffolder-ent2: 61a8d4055db62a1b83fc15e7
      required:
        - _id
...